[#198] Guard PID file creation
Log FATAL when the PID file already exists.

If create_pid() fails because there is already a PID file, a log with
level FATAL is inserted before the program exits.

+++ src/main.c
@@ -1913,8 +1913,24 @@ create_pidfile(void)
 
    config = (struct configuration*)shmem;
 
+   if (strlen(config->pidfile) == 0)
+   {
+      // no pidfile set, use a default one
+      snprintf(config->pidfile, sizeof( config->pidfile ), "%s/pgagraol.%d.pid",
+               config->unix_socket_dir,
+               config->port);
+      pgagroal_log_debug("PID file automatically set to: [%s]", config->pidfile);
+   }
+
    if (strlen(config->pidfile) > 0)
    {
+       // check pidfile is not there
+      if (access(config->pidfile, F_OK) == 0)
+      {
+          pgagroal_log_fatal("PID file [%s] exists, is there another instance running ?", config->pidfile);
+          goto error;
+      }
+
       pid = getpid();
 
       fd = open(config->pidfile, O_WRONLY | O_CREAT | O_EXCL, 0644);
@@ -1949,7 +1965,7 @@ static void remove_pidfile(void)
 
    config = (struct configuration*)shmem;
 
-   if (strlen(config->pidfile) > 0)
+   if (strlen(config->pidfile) > 0 && access(config->pidfile, F_OK) == 0)
    {
       unlink(config->pidfile);
    }
